Split large runtime and keymap tests
